add rescale and convert y to int32 in all datasets

+++ lasagnekit/datasets/cifar10.py
@@ -43,5 +43,5 @@ class Cifar10(object):
             start = end
             end += Cifar10.NB_EXAMPLES_PER_BATCH
 
-        self.X = X
-        self.y = y
+        self.X = X.astype(np.float32)
+        self.y = y.astype(np.int32)
